requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=14.0
twilio>=9.4.0
deepgram-sdk>=3.8.0,<4.0.0
elevenlabs>=1.16.0
openai>=1.60.0
python-dotenv>=1.0.1
httpx>=0.28.0
pydantic>=2.10.0
python-multipart>=0.0.9
